+++ src/AbstractBouncer.php
@@ -13,6 +13,7 @@ use CrowdSec\RemediationEngine\CacheStorage\CacheStorageException;
 use CrowdSec\RemediationEngine\CacheStorage\Memcached;
 use CrowdSec\RemediationEngine\CacheStorage\PhpFiles;
 use CrowdSec\RemediationEngine\CacheStorage\Redis;
+use CrowdSec\RemediationEngine\LapiRemediation;
 use Gregwar\Captcha\CaptchaBuilder;
 use Gregwar\Captcha\PhraseBuilder;
 use IPLib\Factory;
@@ -86,8 +87,26 @@ abstract class AbstractBouncer
         $this->logger->info('Bouncing current IP', [
             'ip' => $ip,
         ]);
-        $remediation = $this->getRemediationForIp($ip);
+        $remediation = $this->getRemediation($ip);
         $this->handleRemediation($remediation, $ip);
+    }
+
+    /**
+     * @throws CacheException
+     * @throws InvalidArgumentException
+     */
+    private function getRemediation(string $ip): string
+    {
+        $remediation = $this->getRemediationForIp($ip);
+        // In case of bypass, we can use AppSec to test also the current request
+        if (
+            $this->getConfig('use_app_sec') && Constants::REMEDIATION_BYPASS === $remediation
+            && $this->remediationEngine instanceof LapiRemediation
+        ) {
+            $remediation = $this->getAppSecRemediationForIp($ip, $this->remediationEngine);
+        }
+
+        return $remediation;
     }
 
     /**
@@ -103,6 +122,23 @@ abstract class AbstractBouncer
             return $this->getRemediationEngine()->clearCache();
         } catch (\Exception $e) {
             throw new BouncerException('Error while clearing cache: ' . $e->getMessage(), (int) $e->getCode(), $e);
+        }
+    }
+
+    /**
+     * Get the remediation for the specified IP using AppSec.
+     *
+     * @throws BouncerException
+     * @throws InvalidArgumentException|CacheException
+     */
+    public function getAppSecRemediationForIp(string $ip, LapiRemediation $remediationEngine): string
+    {
+        try {
+            return $this->capRemediationLevel(
+                $remediationEngine->getAppSecRemediation($this->getAppSecHeaders($ip), $this->getRequestRawBody())
+            );
+        } catch (\Exception $e) {
+            throw new BouncerException($e->getMessage(), (int) $e->getCode(), $e);
         }
     }
 
@@ -174,9 +210,29 @@ abstract class AbstractBouncer
     abstract public function getRemoteIp(): string;
 
     /**
+     * Get current request headers.
+     */
+    abstract public function getRequestHeaders(): array;
+
+    /**
+     * Get current request host.
+     */
+    abstract public function getRequestHost(): string;
+
+    /**
+     * Get current request raw body.
+     */
+    abstract public function getRequestRawBody(): string;
+
+    /**
      * Get current request uri.
      */
     abstract public function getRequestUri(): string;
+
+    /**
+     * Get current request user agent.
+     */
+    abstract public function getRequestUserAgent(): string;
 
     /**
      * This method prune the cache: it removes all the expired cache items.
@@ -287,6 +343,39 @@ abstract class AbstractBouncer
             $message = 'Error while testing cache connection: ' . $e->getMessage();
             throw new BouncerException($message, (int) $e->getCode(), $e);
         }
+    }
+
+    /**
+     * Returns a default "CrowdSec 403" HTML template.
+     * The input $config should match the TemplateConfiguration input format.
+     *
+     * @return string The HTML compiled template
+     */
+    protected function getBanHtml(): string
+    {
+        $template = new Template('ban.html.twig');
+
+        return $template->render($this->configs);
+    }
+
+    /**
+     * Returns a default "CrowdSec Captcha (401)" HTML template.
+     */
+    protected function getCaptchaHtml(
+        bool $error,
+        string $captchaImageSrc,
+        string $captchaResolutionFormUrl
+    ): string {
+        $template = new Template('captcha.html.twig');
+
+        return $template->render(array_merge(
+            $this->configs,
+            [
+                'error' => $error,
+                'captcha_img' => $captchaImageSrc,
+                'captcha_resolution_url' => $captchaResolutionFormUrl,
+            ]
+        ));
     }
 
     /**
@@ -507,42 +596,26 @@ abstract class AbstractBouncer
         $this->sendResponse($body, 401);
     }
 
-    /**
-     * Returns a default "CrowdSec 403" HTML template.
-     * The input $config should match the TemplateConfiguration input format.
-     *
-     * @return string The HTML compiled template
-     */
-    protected function getBanHtml(): string
+    private function getAppSecHeaders(string $ip): array
     {
-        $template = new Template('ban.html.twig');
+        $requestHeaders = $this->getRequestHeaders();
 
-        return $template->render($this->configs);
+        return array_merge(
+            $requestHeaders,
+            [
+                Constants::HEADER_APPSEC_IP => $ip,
+                Constants::HEADER_APPSEC_URI => $this->getRequestUri(),
+                Constants::HEADER_APPSEC_HOST => $this->getRequestHost(),
+                Constants::HEADER_APPSEC_VERB => $this->getHttpMethod(),
+                Constants::HEADER_APPSEC_API_KEY => $this->remediationEngine->getClient()->getConfig('api_key'),
+                Constants::HEADER_APPSEC_USER_AGENT => $this->getRequestUserAgent(),
+            ]
+        );
     }
 
     private function getCache(): AbstractCache
     {
         return $this->getRemediationEngine()->getCacheStorage();
-    }
-
-    /**
-     * Returns a default "CrowdSec Captcha (401)" HTML template.
-     */
-    protected function getCaptchaHtml(
-        bool $error,
-        string $captchaImageSrc,
-        string $captchaResolutionFormUrl
-    ): string {
-        $template = new Template('captcha.html.twig');
-
-        return $template->render(array_merge(
-            $this->configs,
-            [
-                'error' => $error,
-                'captcha_img' => $captchaImageSrc,
-                'captcha_resolution_url' => $captchaResolutionFormUrl,
-            ]
-        ));
     }
 
     /**